segment abstraction
