+++ app/thedailygwei/[ep_number]/components/SegmentsList.tsx
@@ -3,7 +3,7 @@ import { EpisodeProps, SegmentProps } from '@/types';
 
 import { useState, useRef } from 'react';
 
-import EpisodeSegment from '@/app/components/EpisodeSegment';
+import EpisodeSegment from '@/app/thedailygwei/[ep_number]/components/episodeSegment/EpisodeSegment';
 
 interface SegmentsListProps {
   isOrganizedByLength: boolean;
